Delete unused components, move some widgets to components (as they are)
