--- src/widgets/SignInCard.stories.mdx
@@ -1,25 +1,0 @@
-import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import SignInCard from './SignInCard';
-
-<Meta title="Course/SignInCard" component={SignInCard} />
-
-# SignInCard
-
--   was used to show card to sign in to courses
--   barely to not used right now
-
--   use `showTrafficLight` to indicate course participation
--   use `TrafficLightStatus` to differentiate between participation states
--   `tags` use `LFTag` type
--   `data`is of type `LFSubcourse`
-
-<Story name="SignInCard">
-    <SignInCard
-        showTrafficLight
-        trafficLightStatus={'full'}
-        subcourse={{
-            course: { name: 'Test Kurs', outline: 'Lorem Ipsum', description: 'My description', tags: [{ name: 'Englisch' }, { name: 'Deutsch' }] },
-            lectures: [{ start: new Date().toISOString(), duration: 60 }],
-        }}
-    />
-</Story>
