--- src/widgets/StudentCertificates.stories.mdx
@@ -1,12 +1,0 @@
-import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import StudentCertificates from './StudentCertificates';
-
-<Meta title="Misc/StudentCertificates" component={StudentCertificates} />
-
-# StudentCertificates
-
--   loads a student's certificates
-
-<Story name="StudentCertificates">
-    <StudentCertificates filter="" />
-</Story>
