--- src/widgets/UserAchievements.stories.mdx
@@ -1,12 +1,0 @@
-import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import UserAchievements from './UserAchievements';
-
-<Meta title="Misc/UserAchievements" component={UserAchievements} />
-
-# UserAchievements
-
--   pass a local icon to `icon`
-
-<Story name="UserAchievements">
-    <UserAchievements points={20} icon={'icon'} />
-</Story>
