--- src/widgets/WebServiceCard.stories.mdx
@@ -1,20 +1,0 @@
-import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import WebServiceCard from './WebServiceCard';
-
-<Meta title="Misc/WebServiceCard" component={WebServiceCard} />
-
-# WebServiceCard
-
--   used to display external services
-
-<Story name="WebServiceCard">
-    <WebServiceCard
-        title="my Title"
-        description="my description"
-        image="https://picsum.photos/680/480"
-        url="https://google.com"
-        tags={['Informationen', 'PDF']}
-        pros={['Pro1', 'Pro2']}
-        contra={['Contra1']}
-    />
-</Story>
